requirement.txt
flask
flask-cors
librosa
numpy
fpdf
google-genai
soundfile
audioread